+++ src/components/genes/GenesSubscriber.js
@@ -41,12 +41,12 @@ export default function GenesSubscriber(props) {
     dataset,
     geneSelection,
     geneFilter,
+    cellColorEncoding,
   }, {
     setGeneSelection,
     setGeneFilter,
     setGeneHighlight,
     setCellColorEncoding,
-    setCellSetSelection,
   }] = useCoordination(COMPONENT_COORDINATION_TYPES.genes, coordinationScopes);
 
   const [urls, addUrl, resetUrls] = useUrls();
@@ -69,7 +69,6 @@ export default function GenesSubscriber(props) {
   const numGenes = geneList.length;
 
   function setGeneSelectionAndColorEncoding(newSelection) {
-    setCellSetSelection([]);
     setGeneSelection(newSelection);
     setCellColorEncoding('geneSelection');
   }
@@ -88,6 +87,7 @@ export default function GenesSubscriber(props) {
       urls={urls}
     >
       <Genes
+        hasColorEncoding={cellColorEncoding === 'geneSelection'}
         geneList={geneList}
         geneSelection={geneSelection}
         geneFilter={geneFilter}